Dictionary issue fixed

+++ 0x02-python-import_modules/calculator_1.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+
 
 def add(a,b):
     return (a + b)
@@ -7,7 +7,7 @@ def mul(a,b):
     return (a * b)
 
 def div(a,b):
-    return (a / b)
+    return int(a / b)
 
 def sub(a,b):
     return (a - b)
